feat: remake proximityhider
feat: created new main class

+++ orebfuscator-plugin/src/main/java/net/imprex/orebfuscator/config/OrebfuscatorBlockMask.java
@@ -34,7 +34,7 @@ public class OrebfuscatorBlockMask implements BlockMask {
 				blockMask |= BLOCK_MASK_TILEENTITY;
 			}
 
-			this.blockMask[blockId] = (byte) blockMask;
+			this.blockMask[blockId] = (short) blockMask;
 		}
 	}
